added first repository (group) changed data type of ids to Long (former Int) changed application.properties --> no more tracing allowed

+++ group-service/src/main/kotlin/com/placeToBeer/groupService/entities/Membership.kt
@@ -9,7 +9,7 @@ import javax.persistence.*
 data class Membership(
         @Id
         @GeneratedValue(strategy = GenerationType.AUTO)
-        val membershipID: Int,
+        val membershipID: Long,
         @ManyToOne()
         val group: Group,
         @ManyToOne()
